Bug#36178174 Incorrect query result when limit is higher than 4294967294
Changed the internal types inside MRS implementation for variables
representing limit, offset from u32 to u64.
Added tests.

MRS schema still needs update to change the type of items_on_page field.

+++ router/src/mysql_rest_service/include/mrs/database/entry/db_object.h
@@ -57,7 +57,7 @@ class DbObject {
   std::string service_path;
   std::string schema_path;
   std::string object_path;
-  uint32_t on_page;
+  uint64_t on_page;
   std::string db_schema;
   std::string db_table;
   bool requires_authentication;
